+++ docs/supported-libraries.md
@@ -39,7 +39,7 @@ These are the supported libraries and frameworks:
 | [Apache Pulsar](https://pulsar.apache.org/)                                                                                                 | 2.8+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Apache RocketMQ gRPC/Protobuf-based Client](https://rocketmq.apache.org/)                                                                  | 5.0+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Apache RocketMQ Remoting-based Client](https://rocketmq.apache.org/)                                                                       | 4.8+                               | [opentelemetry-rocketmq-client-4.8](../instrumentation/rocketmq/rocketmq-client/rocketmq-client-4.8/library)                                                                                                                                                                                                                                                                            | [Messaging Spans]                                                                                                                       |
-| [Apache Struts 2](https://github.com/apache/struts)                                                                                         | 2.3+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | Provides `http.route` [2], Controller Spans [3]                                                                                         |
+| [Apache Struts](https://github.com/apache/struts)                                                                                           | 2.3+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | Provides `http.route` [2], Controller Spans [3]                                                                                         |
 | [Apache Tapestry](https://tapestry.apache.org/)                                                                                             | 5.4+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | Provides `http.route` [2], Controller Spans [3]                                                                                         |
 | [Apache Wicket](https://wicket.apache.org/)                                                                                                 | 8.0+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | Provides `http.route` [2]                                                                                                               |
 | [Armeria](https://armeria.dev)                                                                                                              | 1.3+                               | [opentelemetry-armeria-1.3](../instrumentation/armeria/armeria-1.3/library)                                                                                                                                                                                                                                                                                                             | [HTTP Client Spans], [HTTP Client Metrics], [HTTP Server Spans], [HTTP Server Metrics]                                                  |
