+++ instrumentation/struts/struts-7.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/struts/v7_0/StrutsCodeAttributesGetter.java
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.struts2.v2_3;
+package io.opentelemetry.javaagent.instrumentation.struts.v7_0;
 
-import com.opensymphony.xwork2.ActionInvocation;
 import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesGetter;
+import org.apache.struts2.ActionInvocation;
 
 public class StrutsCodeAttributesGetter implements CodeAttributesGetter<ActionInvocation> {
